+++ src/mcp/tools/dataconnect/execute_graphql.ts
@@ -40,12 +40,12 @@ export const execute_graphql = tool(
     { query, service_id, variables: unparsedVariables, use_emulator },
     { projectId, config, host },
   ) => {
-    const serviceInfo = await pickService(projectId!, config!, service_id || undefined);
+    const serviceInfo = await pickService(projectId, config, service_id || undefined);
 
     let apiClient: Client;
 
     if (use_emulator) {
-      apiClient = await getDataConnectEmulatorClient(await host.getEmulatorHubClient());
+      apiClient = await getDataConnectEmulatorClient(host);
     } else {
       apiClient = dataplane.dataconnectDataplaneClient();
     }