+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 - Fixed bug where functions packaged as ESM doesn't load on emulator for some versions of Node.js 22. (#8394)
+- Updated the Data Connect emulator to use pglite 0.3.x and Postgres 17, which fixes some crashes related to wire protocol inconsistencies. (#8679, #8658)
+- Remove container cleanup logic in functions:delete command (#8771)
+- Fixed an issue where the IAM enablement for GenKit monitoring would try to change an invalid service account. (#8756)
+- Added caching to API enablement checks to reduce burn of `serviceusage.googleapis.com` quota.